refactor step 2
